Refactor Builders
