+++ googlecast/src/main/java/com/kaltura/playkit/plugins/googlecast/caf/KalturaCastBuilder.java
@@ -27,40 +27,6 @@ import static com.kaltura.playkit.plugins.googlecast.caf.KalturaCastInfo.MEDIA_I
 import static com.kaltura.playkit.plugins.googlecast.caf.KalturaCastInfo.TEXT_LANGUAGE;
 import static com.kaltura.playkit.plugins.googlecast.caf.KalturaCastInfo.VMAP_ADS_REQUEST;
 
-public class KalturaCastBuilder extends CAFCastBuilder {
+public class KalturaCastBuilder extends CAFCastBuilder<KalturaCastBuilder> {
 
-    public KalturaCastBuilder() {
-        castInfo = new KalturaCastInfo();
-    }
-
-    public JSONObject getCustomData() {
-        JSONObject customData = new JSONObject();
-        try {
-            JSONObject mediaData = new JSONObject();
-            mediaData.put(ENTRY_ID, castInfo.getMediaEntryId());
-            if (!TextUtils.isEmpty(castInfo.getKs())) {
-                mediaData.put(KS, castInfo.getKs());
-            }
-
-            customData.put(MEDIA_INFO, mediaData);
-            if (!TextUtils.isEmpty(castInfo.getTextLanguage())) {
-                customData.put(TEXT_LANGUAGE, castInfo.getTextLanguage());
-            }
-            if (!TextUtils.isEmpty(castInfo.getAudioLanguage())) {
-                customData.put(AUDIO_LANGUAGE, castInfo.getAudioLanguage());
-            }
-
-            if(castInfo.getAdsModel() != null && castInfo.getAdsModel().getAdTagType() == CAFCastBuilder.AdTagType.VMAP) {
-                if (((VmapAdsModel)castInfo.getAdsModel()).getVmapAdRequest() != null) {
-                    JSONObject vmapAdRequestJson = ((VmapAdsModel)castInfo.getAdsModel()).getVmapAdRequest().toJSONObject();
-                    if (vmapAdRequestJson != null) {
-                        customData.put(VMAP_ADS_REQUEST, vmapAdRequestJson);
-                    }
-                }
-            }
-        } catch (JSONException e) {
-            e.printStackTrace();
-        }
-        return customData;
-    }
 }
